refactor(api): extract models, use string for ip and method fields, rework router
