+++ api/lib/rkwst_web/router.ex
@@ -7,6 +7,14 @@ defmodule RkwstWeb.Router do
 
   scope "/api", RkwstWeb do
     pipe_through :api
+
+    scope "/v1", as: :v1 do
+      get "/bins/:id", BinController, :show
+      put "/bins/:id", BinController, :update
+      post "/bins", BinController, :create
+
+      get "/bins/:id/requests", RequestController, :show
+    end
   end
 
   # Enables LiveDashboard only for development
@@ -19,18 +27,13 @@ defmodule RkwstWeb.Router do
   if Mix.env() in [:dev, :test] do
     import Phoenix.LiveDashboard.Router
 
-    scope "/api", RkwstWeb, as: :api do
+    scope "/api", RkwstWeb do
       pipe_through :api
 
       scope "/v1", as: :v1 do
         get "/bins", BinController, :index
-        get "/bins/:id", BinController, :show
-        put "/bins/:id", BinController, :update
-        post "/bins", BinController, :create
-
-        get "/requests", RequestController, :index # for debug
-        get "/bins/:id/requests", RequestController, :show
-        post "/bins/:id/requests", RequestController, :create # for debug
+        get "/requests", RequestController, :index
+        post "/bins/:id/requests", RequestController, :create
       end
     end
